+++ test/unit/compiler/concerns/compiles_components_test.rb
@@ -154,4 +154,19 @@ class CompilesComponentsTest < TestCase
     end
     assert_equal "Unexpected closing tag </x-button>, expecting </x-link>", exception.to_s
   end
+
+  def test_partial_assigns_locals
+    assert_partial_compiles_to "@props(cake: required){{ cake }}", "choccy", locals: {cake: 'choccy'}
+
+    # When enabled, components rendered normally should still work
+    assert_partial_compiles_to "<x-simple_button label='choccy'/>", "<button>choccy</button>", locals: {cake: 'not choccy'}
+  end
+
+  def test_partial_assigns_slot
+    assert_partial_compiles_to "{{ slot }}", "choccy" do "choccy" end
+    assert_partial_compiles_to "{{ slot }}", "choccy", locals: {slot: 'choccy'}
+
+    # When enabled, components rendered normally should still work
+    assert_partial_compiles_to "<x-button>choccy<//>", "<button class=\"button\">choccy</button>" do "not choccy" end
+  end
 end