+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.1.0 [2025-03-14]
 - Add ability to add raw directive handlers that add ruby code to the template
+- Add `RBlade.direct_component_rendering` option to allow RBlade components to be rendered directly
 - Improve performance of regular expressions
 - Update README
 - Fix issue with component tokenizer hanging